% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorRamp.R
\name{colorRampPalette}
\alias{colorRampPalette}
\title{Color ramp palette}
\usage{
colorRampPalette(colors, ...)
}
\arguments{
\item{colors}{palette used}

\item{...}{other arguments sent to grDevices::colorRamp()}
}
\value{
a color palette compatible with ggplot
}
\description{
This function uses the selected palette if length(pal) < n, where n is
the number of groups in fill/color used in scale_*_pwc_d/c
}
\details{
If length(pal) > n, it will create new colors based on grDevices::colorRamp()

Most of the code here is from the function grDevices::colorRampPalette, but with some tweaks
to fit this purpose.
}
