% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cols.R
\name{pwc_cols}
\alias{pwc_cols}
\title{Funksjon for å hente hex-kode til farger}
\usage{
pwc_cols(...)
}
\arguments{
\item{...}{Navn på farger i listen pwc_colors}
}
\description{
Funksjon for å hente hex-kode til farger
}
