% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pals.R
\name{pwc_pal}
\alias{pwc_pal}
\title{Fargepaletter for pwc}
\usage{
pwc_pal(palette = "pwc_palette", rev = FALSE, for_print = FALSE, ...)
}
\arguments{
\item{palette}{Navn på palette fra listen pwc_pal}

\item{rev}{Logisk variabel som indikerer om paletten skal reverseres eller ikke}

\item{...}{Ytterlige variabler som sendes til colorRampPalette()}
}
\value{
Objekt med en palett
}
\description{
#' Returns a color or a palette
}
