% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_funcs.R
\name{scale_color_pwc}
\alias{scale_color_pwc}
\title{Scale-funksjon for pwc-farger}
\usage{
scale_color_pwc(palette = "pwc_palette", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Navn på palett fra listen pwc_pal}

\item{discrete}{Logisk variabel som indikerer om farge-aestichen er diskret eller ikke}

\item{reverse}{Logisk variabel som indikerer om paletten skal reverseres eller ikke.}

\item{...}{Flere argumenter som sendes til hhv discrete_scale og scale_fill_gradientn for diskret TRUE/FALSE}
}
\description{
Scale-funksjon for pwc-farger
}
