% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_funcs.R
\name{scale_fill_pwc_d}
\alias{scale_fill_pwc_d}
\title{scale_fill_pwc_d}
\usage{
scale_fill_pwc_d(palette = "pwc_palette", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Navn på palett fra listen pwc_pal}

\item{reverse}{Logisk variabel som indikerer om paletten skal reverseres eller ikke.}

\item{...}{Flere argumenter som sendes til hhv discrete_scale og scale_fill_gradientn for diskret TRUE/FALSE}
}
\value{

}
\description{
scale_fill_pwc_d
}
