% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pwc.R
\name{theme_pwc}
\alias{theme_pwc}
\title{Theme PwC}
\usage{
theme_pwc(base_size = 2.5, font_fam = "Arial")
}
\arguments{
\item{font_fam}{Skrifttype, f.eks. Georgia eller Arial.}
}
\value{

}
\description{
Theme PwC
}
